refactor: 당첨 로또 만드는 폼, 당첨 통계 모달 만드는 폼 분리
